MIM-170 拆分照片权限测试以通过源码体积门禁
