+++ macos/MimiRemoteMac/Tests/HostStoreTests.swift
@@ -4,76 +4,6 @@ import XCTest
 
 @MainActor
 final class HostStoreTests: XCTestCase {
-    func testBootstrapRequestsPhotoAuthorizationBeforeRegisteringAgent() async {
-        let events = EventRecorder()
-        var authorization = PhotoLibraryAuthorization.notDetermined
-        let store = makeStore(
-            configExists: true,
-            registerAgent: { events.append("register-mac") },
-            photoLibraryAccess: PhotoLibraryAccessClient(
-                authorizationStatus: { authorization },
-                requestAuthorization: {
-                    events.append("request-photos")
-                    authorization = .authorized
-                    return authorization
-                },
-                openPhotosPrivacySettings: {},
-                openFullDiskAccessSettings: {}
-            )
-        )
-
-        await store.bootstrap()
-
-        XCTAssertEqual(events.values, ["request-photos", "register-mac"])
-        XCTAssertEqual(store.photoLibraryAuthorization, .authorized)
-        XCTAssertEqual(store.lifecycle, .ready)
-    }
-
-    func testDeniedPhotoAuthorizationDoesNotBlockAgentStartup() async {
-        let events = EventRecorder()
-        let store = makeStore(
-            configExists: true,
-            registerAgent: { events.append("register-mac") },
-            photoLibraryAccess: PhotoLibraryAccessClient(
-                authorizationStatus: { .notDetermined },
-                requestAuthorization: {
-                    events.append("request-photos")
-                    return .denied
-                },
-                openPhotosPrivacySettings: {},
-                openFullDiskAccessSettings: {}
-            )
-        )
-
-        await store.bootstrap()
-
-        XCTAssertEqual(events.values, ["request-photos", "register-mac"])
-        XCTAssertEqual(store.photoLibraryAuthorization, .denied)
-        XCTAssertEqual(store.lifecycle, .ready)
-        XCTAssertNil(store.lastError)
-    }
-
-    func testBootstrapWithoutConfigurationDoesNotRequestPhotoAuthorization() async {
-        let events = EventRecorder()
-        let store = makeStore(
-            configExists: false,
-            photoLibraryAccess: PhotoLibraryAccessClient(
-                authorizationStatus: { .notDetermined },
-                requestAuthorization: {
-                    events.append("request-photos")
-                    return .authorized
-                },
-                openPhotosPrivacySettings: {},
-                openFullDiskAccessSettings: {}
-            )
-        )
-
-        await store.bootstrap()
-
-        XCTAssertTrue(events.values.isEmpty)
-        XCTAssertEqual(store.lifecycle, .notConfigured)
-    }
-
     func testBootstrapRegistersBundledAgentWhenServiceRecordIsNotFound() async {
         let events = EventRecorder()
         var registrationState = ServiceRegistrationState.notFound
@@ -2147,7 +2077,7 @@ final class HostStoreTests: XCTestCase {
         XCTAssertTrue(backendCalls.values.isEmpty)
     }
 
-    private func makeStore(
+    func makeStore(
         configExists: Bool,
         homebrewLoaded: Bool = false,
         agentStatus: @escaping @MainActor () -> ServiceRegistrationState = { .notRegistered },
@@ -2408,7 +2338,7 @@ private enum TestError: LocalizedError {
     var errorDescription: String? { "预期的测试错误" }
 }
 
-private final class EventRecorder: @unchecked Sendable {
+final class EventRecorder: @unchecked Sendable {
     private let lock = NSLock()
     private var storage: [String] = []
